* [iOS] fix build error

+++ ios/sdk/WeexSDK/Sources/Layout/WXComponent+Layout.mm
@@ -761,7 +761,7 @@ static WeexCore::WXCoreSize flexCssNodeMeasure(WeexCore::WXCoreLayoutNode *node,
         NSLog(@"test -> _insertChildCssNode [%@<-%@],index :[%ld,%ld]",
               self.type,
               subcomponent.type,
-              index,
+              (long)index,
               self.flexCssNode->getChildCount()
               );
     }
